requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
bleach==5.0.1
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.50.0
django-annoying==0.10.6
django-cors-headers==3.13.0
django-post-office==3.6.3
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
jsonfield==3.1.0
Markdown==3.4.1
msgpack==1.0.4
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
selenium==4.6.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tinycss2==1.1.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
wrapt==1.14.1
wsproto==1.2.0
zope.interface==5.5.1
